refactoring - visualize after dataset change
