+++ src/components/PathBar.vue
@@ -1,0 +1,18 @@
+<template>
+  <div></div>
+</template>
+
+<script lang="ts">
+import Vue from 'vue'
+
+export default Vue.extend({
+  name: 'PathBar',
+
+  data: () => ({
+  }),
+  computed: {
+  },
+  methods: {
+  }
+})
+</script>
