+++ src/models/ArrowData.ts
@@ -1,0 +1,5 @@
+export interface ArrowData {
+    id: string;
+    label: string;
+    word?: string;
+}
