+++ src/models/ComboboxItem.ts
@@ -1,0 +1,9 @@
+export class ComboboxItem {
+    id: number;
+    name: string;
+
+    constructor (name: string, id: number) {
+      this.name = name
+      this.id = id
+    }
+}
