+++ src/models/MappingData.ts
@@ -1,0 +1,6 @@
+export interface MappingData {
+    id: string;
+    group: string;
+    shared: number;
+    size: number;
+}
