+++ src/models/MappingNode.ts
@@ -1,0 +1,7 @@
+export interface MappingNode {
+    id: number;
+    nodeID?: string;
+    mapBy?: string;
+    name: string;
+    children?: MappingNode[];
+}
