+++ src/models/PathNode.ts
@@ -1,0 +1,4 @@
+export interface PathNode {
+    id: number;
+    name: string;
+}
