+++ src/models/Position.ts
@@ -1,0 +1,3 @@
+export enum Position {
+    Left, Right
+}
